// backend/src/modules/users/enums/success.enum.ts
export enum ESuccess {
  CREATE_USER = 'Usuário cadastrado com sucesso!',
  USER_FOUND = 'Usuário encontrado com sucesso!',
  USERS_FOUND = 'Usuários encontrado com sucesso!',

  PASSWORD_UPDATE = 'Senha de usuário atualizada com sucesso!',
  ADMIN_UPDATE = 'Nível de acesso alterado com sucesso!',

  DELETE_USER = 'Usuário deletado com sucesso!',
}